+++ src/Hagar.CodeGenerator/InvokableGenerator.cs
@@ -592,7 +592,7 @@ namespace Hagar.CodeGenerator
         {
             var fields = new List<FieldDescription>();
 
-            uint fieldId = 0;
+            ushort fieldId = 0;
             foreach (var parameter in method.Parameters)
             {
                 fields.Add(new MethodParameterFieldDescription(parameter, $"arg{fieldId}", fieldId));
@@ -678,7 +678,7 @@ namespace Hagar.CodeGenerator
 
         internal class MethodParameterFieldDescription : FieldDescription, IMemberDescription
         {
-            public MethodParameterFieldDescription(IParameterSymbol parameter, string fieldName, uint fieldId)
+            public MethodParameterFieldDescription(IParameterSymbol parameter, string fieldName, ushort fieldId)
                 : base(parameter.Type, fieldName)
             {
                 FieldId = fieldId;
@@ -688,7 +688,7 @@ namespace Hagar.CodeGenerator
             public int ParameterOrdinal => Parameter.Ordinal;
 
             public override bool IsInjected => false;
-            public uint FieldId { get; }
+            public ushort FieldId { get; }
             public ISymbol Member => Parameter;
             public ITypeSymbol Type => FieldType;
             public IParameterSymbol Parameter { get; }
